chore(seismic-bot): move shared faucet types under src/
`Key`, `FaucetConfig`, and `Faucets` lived in the package's root
`index.ts`, which is unreachable from `src/` via the `@sbot/*` alias.
ESLint's `no-relative-import-paths` kept rewriting `'../index.ts'` to
`'packages/seismic-bot/index.ts'` — a workspace-rooted specifier
NodeNext can't resolve.

Move the types into `src/types.ts` so both `index.ts` and the rest of
the package import them via the existing `@sbot/types.ts` alias.
